Renamed PopupService.ts to popup.service.ts
